+++ CHANGELOG.md
@@ -10,6 +10,9 @@ All notable changes to this project will be documented in this file.
 - Added `json.stringify` handler.
 - Added `net.fetch` handler.
 - Added `os.platform` handler.
+- Added `forEach` handler.
+- Added `forIn` handler.
+- Added `forOf` handler.
 
 ### Fixed
 - Fixed a typo in the error message shown from the help command when the user tries to show details about a non-existent command.